third_party/WebKit/LayoutTests/fast/events/press-enter-key-on-focused-element-inside-link.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<a id="link" href="#" tabindex="-1" onclick="debug(eventInfo(event))">
    <h1 tabindex="1" id="focus0"
        onkeydown="debug(eventInfo(event))"
        onkeypress="debug(eventInfo(event))"
        onkeyup="debug(eventInfo(event))">EnterMe</h1>
    <textarea tabindex="2" id="focus1"
        onkeydown="debug(eventInfo(event))"
        onkeypress="debug(eventInfo(event))"
        onkeyup="debug(eventInfo(event))">TypeHere</textarea>
</a>
<svg width="200" height="50" onclick="debug(eventInfo(event))">
    <a xlink:href="#">
        <text font-size="18px" y="20" tabindex="3"
        onkeydown="debug(eventInfo(event))"
        onkeypress="debug(eventInfo(event))"
        onkeyup="debug(eventInfo(event))">SVG anchor with tabindex</text>
    </a>
</svg>
<div id="console"></div>
<script>
    description('This test ensures that the link is activated by enter key on focusing the child element inside the anchor');
    function eventInfo(event) {
        if (event.type == 'keydown' || event.type == 'keypress' || event.type == 'keyup')
            return event.keyIdentifier == 'U+0009' ? '' : event.target.tagName + ' - ' + event.type + ' - ' + event.keyIdentifier;
        else if (event.type == 'click')
            return event.target.tagName + ' - ' + event.type;
        else
            return 'unexpected event - ' + event.target.tagName + ' - ' + event.type;

    }
    for (var i = 0; i < 3; i++) { 
        if (window.eventSender) {
            eventSender.keyDown('\t');
            eventSender.keyDown('\n');
        }
    }

</script>
